+++ src/patos/__init__.py
@@ -4,7 +4,7 @@ from .dispatch import value_dispatch
 from .flyweight import FlyweightMeta
 from .registry import Registry
 from .singleton import Singleton, SingletonMeta
-from .strategy import Strategy
+from .strategy import Available, Strategy, StrategyError
 from .strflag import StrFlag
 
 try:
@@ -13,11 +13,13 @@ except PackageNotFoundError:
     __version__ = "0.0.0"
 
 __all__ = [
+    "Available",
     "FlyweightMeta",
     "Registry",
     "Singleton",
     "SingletonMeta",
     "StrFlag",
     "Strategy",
+    "StrategyError",
     "value_dispatch",
 ]